api/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-cors-headers==3.10.1
django-environ==0.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==19.9.0